--- apps/admin/src/components/TextField/index.tsx
@@ -1,2 +1,0 @@
-export * from './RemovableTextField';
-export * from './AdminTextField';
